+++ lib/src/styles/theme.dart
@@ -176,12 +176,6 @@ class ThemeData with Diagnosticable {
   final Duration slowAnimationDuration;
   final Curve animationCurve;
 
-  /// The mouse cursor used by many inputs, such as [Button],
-  /// [RadioButton] and [ToggleSwitch]. By default, if the
-  /// state is pressing or hovering, [SystemMouseCursors.click]
-  /// is used, otherwise [MouseCursor.defer] is used.
-  final ButtonState<MouseCursor> inputMouseCursor;
-
   final Brightness brightness;
   final VisualDensity visualDensity;
 
@@ -242,7 +236,6 @@ class ThemeData with Diagnosticable {
     required this.infoBarTheme,
     required this.focusTheme,
     required this.scrollbarTheme,
-    required this.inputMouseCursor,
     required this.snackbarTheme,
     required this.pillButtonBarTheme,
     required this.bottomSheetTheme,
@@ -270,7 +263,6 @@ class ThemeData with Diagnosticable {
     Color? acrylicBackgroundColor,
     Color? micaBackgroundColor,
     Color? shadowColor,
-    ButtonState<MouseCursor>? inputMouseCursor,
     Duration? fasterAnimationDuration,
     Duration? fastAnimationDuration,
     Duration? mediumAnimationDuration,
@@ -325,12 +317,6 @@ class ThemeData with Diagnosticable {
     typography = Typography.standard(brightness: brightness)
         .merge(typography)
         .apply(fontFamily: fontFamily);
-    inputMouseCursor ??= ButtonState.resolveWith((states) {
-      // if (states.isHovering || states.isPressing) {
-      //   return SystemMouseCursors.click;
-      // }
-      return MouseCursor.defer;
-    });
     focusTheme = FocusThemeData.standard(
       glowColor: accentColor.withOpacity(0.15),
       primaryBorderColor: inactiveColor,
@@ -354,7 +340,6 @@ class ThemeData with Diagnosticable {
       backgroundColor: micaBackgroundColor,
       disabledColor: disabledColor,
       highlightColor: accentColor,
-      inputMouseCursor: inputMouseCursor,
       typography: typography,
       inactiveColor: inactiveColor,
     );
@@ -401,7 +386,6 @@ class ThemeData with Diagnosticable {
       toggleSwitchTheme: toggleSwitchTheme,
       tooltipTheme: tooltipTheme,
       typography: typography,
-      inputMouseCursor: inputMouseCursor,
       snackbarTheme: snackbarTheme,
       pillButtonBarTheme: pillButtonBarTheme,
       bottomSheetTheme: bottomSheetTheme,
@@ -435,7 +419,6 @@ class ThemeData with Diagnosticable {
       slowAnimationDuration:
           lerpDuration(a.slowAnimationDuration, b.slowAnimationDuration, t),
       animationCurve: t < 0.5 ? a.animationCurve : b.animationCurve,
-      inputMouseCursor: t < 0.5 ? a.inputMouseCursor : b.inputMouseCursor,
       buttonTheme: ButtonThemeData.lerp(a.buttonTheme, b.buttonTheme, t),
       checkboxTheme:
           CheckboxThemeData.lerp(a.checkboxTheme, b.checkboxTheme, t),
@@ -488,7 +471,6 @@ class ThemeData with Diagnosticable {
     Duration? mediumAnimationDuration,
     Duration? slowAnimationDuration,
     Curve? animationCurve,
-    ButtonState<MouseCursor>? inputMouseCursor,
     ButtonThemeData? buttonTheme,
     BottomNavigationThemeData? bottomNavigationTheme,
     BottomSheetThemeData? bottomSheetTheme,
@@ -534,7 +516,6 @@ class ThemeData with Diagnosticable {
           mediumAnimationDuration ?? this.mediumAnimationDuration,
       slowAnimationDuration:
           slowAnimationDuration ?? this.slowAnimationDuration,
-      inputMouseCursor: inputMouseCursor ?? this.inputMouseCursor,
       animationCurve: animationCurve ?? this.animationCurve,
       buttonTheme: this.buttonTheme.merge(buttonTheme),
       bottomNavigationTheme: